+++ app-code/spring-cloud-resilience4j-webflux/src/main/java/com/csoftz/spring/cloud/resilience4j/webflux/common/config/CircuitBreakerConfiguration.java
@@ -1,0 +1,34 @@
+package com.csoftz.spring.cloud.resilience4j.webflux.common.config;
+
+import java.time.Duration;
+
+import org.springframework.cloud.circuitbreaker.resilience4j.ReactiveResilience4JCircuitBreakerFactory;
+import org.springframework.cloud.circuitbreaker.resilience4j.Resilience4JConfigBuilder;
+import org.springframework.cloud.client.circuitbreaker.Customizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import io.github.resilience4j.circuitbreaker.CircuitBreakerConfig;
+import io.github.resilience4j.timelimiter.TimeLimiterConfig;
+
+@Configuration
+public class CircuitBreakerConfiguration {
+
+    @Bean
+    public Customizer<ReactiveResilience4JCircuitBreakerFactory> defaultCustomizer() {
+        return factory ->
+            factory.configureDefault(id -> new Resilience4JConfigBuilder(id)
+                .circuitBreakerConfig(CircuitBreakerConfig.ofDefaults())
+                .timeLimiterConfig(TimeLimiterConfig.custom().timeoutDuration(Duration.ofSeconds(4)).build())
+                .build());
+    }
+
+    @Bean
+    public Customizer<ReactiveResilience4JCircuitBreakerFactory> customerServiceCustomizer() {
+        return factory ->
+            factory.configure(builder ->
+                builder
+                    .timeLimiterConfig(TimeLimiterConfig.custom().timeoutDuration(Duration.ofSeconds(2)).build())
+                    .circuitBreakerConfig(CircuitBreakerConfig.ofDefaults()), "customer-service");
+    }
+}