+++ app-code/spring-cloud-resilience4j-webflux/src/main/java/com/csoftz/spring/cloud/resilience4j/webflux/common/config/WebClientConfiguration.java
@@ -1,0 +1,19 @@
+package com.csoftz.spring.cloud.resilience4j.webflux.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfiguration {
+
+    @Bean
+    public WebClient webClient() {
+        return WebClient.builder()
+            .baseUrl("http://localhost:8500")
+            .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+            .build();
+    }
+}