+++ app-code/spring-cloud-resilience4j-webflux/src/test/java/com/csoftz/spring/cloud/resilience4j/webflux/ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.csoftz.spring.cloud.resilience4j.webflux;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}