+++ app-code/spring-cloud-resilience4j-webmvc/src/main/java/com/csoftz/spring/cloud/resilience4j/webmvc/common/exception/NotFoundException.java
@@ -1,0 +1,27 @@
+/*----------------------------------------------------------------------------*/
+/* Source File:   NOTFOUNDEXCEPTION.JAVA                                      */
+/* Copyright (c), 2022 CSoftZ                                                 */
+/*----------------------------------------------------------------------------*/
+/*-----------------------------------------------------------------------------
+ History
+ Jun.09/2022  COQ  File created.
+ -----------------------------------------------------------------------------*/
+package com.csoftz.spring.cloud.resilience4j.webmvc.common.exception;
+
+public class NotFoundException extends Exception {
+    public NotFoundException() {
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+
+    public NotFoundException(Throwable cause) {
+        super(cause);
+    }
+
+    @Override
+    public String toString() {
+        return "NotFoundException: " + getMessage();
+    }
+}